Updated my tkinter file (characterinterface.py)
I was crashing and generating memory leaks with my old GUI, which simply overwrote all the panels with each refresh (durp!).  I converted the tkinter loop to a class object and modified it to update only the text and button widgets upon each refresh.  Took me a month to restore all the functionality I had in the old version, but I'm finally back to making forward progress as of today's update.