Buttoned up the Bulk Generator
Pretty narrow update, but a satisfying one. Refactored many of the bulk update routines. Also added additional hidden buttons for output navigation. The updates to generate character are exclusively to the to-do list.

Now I should be entering into the home stretch. I've got a handful of punchlist items to work through (minor quality of life items) and then it's on to .exe generation. Hopefully it won't be another ten month gap!